% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcd.R
\name{bcd}
\alias{bcd}
\title{Layer contribution to diversity}
\usage{
bcd(x, percentage=TRUE)
}
\arguments{
\item{x}{filename (character), Raster* object}

\item{percentage}{logical. If TRUE the contribution of each layer is expressed as a percentage.}
}
\value{
Named numeric vector with the contribution of each layer to diversity
}
\description{
bcd computes the contribution of each layer to the overall diversity as proposed by Rossi et al. (2021). The diversity
calculation is based on the sum of squares divided by the number of pixels and layers. The diversity accounts for the pairwise
dissimilarity between pixels belonging to the same layer as well as the dissimilarity between pixels of
different layers. All pixels are considered equally important.
}
\examples{
\dontrun{#Dataset contribution to diversity
ex <-bcd(MTCI.stabelchod)
# MTCI.stabelchod is a RasterStack containing
# a time series of spectral information
}

}
\references{
Rossi, C., Kneubühler, M., Schütz, M., Schaepman, M.E, Haller, R.M., & Risch, A.C. (2021). Ecological Indicators
(\href{https://doi.org/10.1016/j.ecolind.2021.108106})
}
\author{
Christian Rossi christian.rossi1990@gmail.com
}
