% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcd.R
\name{lcd}
\alias{lcd}
\title{Local contribution of each pixel to diversity}
\usage{
lcd(x, percentage=TRUE)
}
\arguments{
\item{x}{filename (character), Raster* object}

\item{percentage}{logical. If TRUE the contribution of each pixel is expressed as a percentage.}
}
\value{
Raster* object with the local contribution of each pixel to diversity
}
\description{
lcd computes the contribution of each pixel to the overall diversity as proposed by Rossi et al. (2021). The diversity
calculation is based on the sum of squares divided by the number of pixels and layers. The diversity accounts for the pairwise
dissimilarity between pixels belonging to the same layer as well as the dissimilarity between pixels of
different layers. All pixels are considered equally important.
}
\examples{
\dontrun{#plot pixel contribution to diversity
plot(lcd(MTCI.stabelchod))
# MTCI.stabelchod is a RasterStack containing
# a time series of spectral information
}

}
\references{
Rossi, C., Kneubühler, M., Schütz, M., Schaepman, M.E, Haller, R.M., & Risch, A.C. (2021). Remote sensing of spectral diversity:
A new methodological approach to account for spatio-temporal dissimilarities between plant communities. Ecological Indicators, 130, 108106.
}
\author{
Christian Rossi christian.rossi1990@gmail.com
}
