% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCI.stabelchod.R
\docType{data}
\name{MTCI.stabelchod}
\alias{MTCI.stabelchod}
\title{MTCI time series of Stabelchod in the Swiss National Park}
\format{
An object of class RasterStack.
}
\usage{
data(MTCI.stabelchod)
}
\description{
MTCI (MERIS terrestrial chlorophyll index) was calculated from eleven atmospherically
corrected Sentinel-2 images from summer 2017 and 2018. Fo Further details see Rossi et al., (2021).
The extent of the raster is constrained to the meadow in Stabelchod and the surrounding forest.
}
\references{
Rossi, C., Kneubühler, M., Schütz, M., Schaepman, M.E, Haller, R.M., & Risch, A.C. (2021). Remote sensing of spectral diversity:
A new methodological approach to account for spatio-temporal dissimilarities between plant communities. Ecological Indicators, 130, 108106.
(\href{https://doi.org/10.1016/j.ecolind.2021.108106})
}
\keyword{datasets}
