% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcd.R
\name{bcd}
\alias{bcd}
\title{Layer contribution to diversity}
\usage{
typical usage is

bcd(x)

where x is a RasterStack containing a time series of one spectral information (e.g., NDVI) or
plant trait (e.g., specific leaf area).
}
\arguments{
\item{x}{filename (character), Raster* object}

\item{percentage}{logical. If TRUE the contribution of each layer is expressed as percentage.}
}
\value{
Named numeric vector with contribution of each layer to diversity
}
\description{
bcd computes the contribution of each layer to the overall diversity as proposed by Rossi et al. (2020). The diversity
calculation is based on the sum of square divided by the number of pixels and layers. The diversity accounts for the pairwise
dissimilarity between pixels belonging to the same layer as well as the dissimilarity between pixels of
different layers. All pixels are considered equally important.
}
\examples{
#Dataset contribution to diversity
ex <-bcd(MTCI.stabelchod)



}
\references{
Rossi et al. (2020) Remote Sensing of Environment 236:111415
(\href{https://www.sciencedirect.com/science/article/pii/S0034425719304341}{ScienceDirect})
}
\author{
Christian Rossi christian.rossi1990@gmail.com
}
