% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divcom.R
\name{divcom}
\alias{divcom}
\title{Total diversity and its components: spatial, layer (temporal) and interaction term}
\usage{
typical usage is

divcom(x)

where x is a RasterStack containing a time series of one spectral information (e.g., NDVI) or
plant trait (e.g., specific leaf area).
}
\arguments{
\item{x}{filename (character), Raster* object}
}
\value{
Named numeric vector with contribution of each layer to diversity
}
\description{
divcom partitions the overall diversity in a spatial, temporal and interaction term as proposed by Rossi et al. (2020).
The diversity calculation is based on the sum of square divided by the number of pixels and layers. The diversity accounts for
for the pairwise dissimilarity between pixels belonging to the same layer as well as the dissimilarity between pixels
of different layers. All pixels are considered equally important.
}
\examples{
#Calculate total diverstiy and its components of a RasterStack
ex <-divcom(MTCI.stabelchod)

}
\references{
Rossi et al. (2020) Remote Sensing of Environment 236:111415
(\href{https://www.sciencedirect.com/science/article/pii/S0034425719304341}{ScienceDirect})
}
\author{
Christian Rossi christian.rossi1990@gmail.com
}
