% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCI.stabelchod.R
\docType{data}
\name{MTCI.stabelchod}
\alias{MTCI.stabelchod}
\title{MTCI time series of Stabelchod in the Swiss National Park}
\format{An object of class RasterStack.}
\usage{
data(MTCI.stabelchod)
}
\description{
MTCI (MERIS terrestrial chlorophyll index) was calculated from eleven atmospherically
corrected Sentinel-2 images from summer 2017 and 2018. Fo Further details see Rossi et al., (2020).
The extent of the raster is constrained to the meadow in Stabelchod and the surrounding forest.
}
\references{
Rossi et al. (2020) Remote Sensing of Environment 236:111415
(\href{https://www.sciencedirect.com/science/article/pii/S0034425719304341}{ScienceDirect})
}
\keyword{datasets}
